+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [#1402](https://github.com/hyperf/hyperf/pull/1402) Added `Hyperf\DbConnection\Annotation\Transactional` to begin a transaction automatically.
 - [#1412](https://github.com/hyperf/hyperf/pull/1412) Added `Hyperf\View\RenderInterface::getContents()`.
+- [#1416](https://github.com/hyperf/hyperf/pull/1416) Added server event constants `ON_WORKER_ERROR`.
 
 ## Fixed